change sql types of money and percentage to double precision and real respectively

+++ accounting_system_rust/accounting_system/src/invoicing/additional_charge/additional_charge_sql/additional_charge_ddl.sql
@@ -5,7 +5,7 @@ create table additional_charge
     invoice_table_id uuid references invoice (id),
     line_no          smallint                        not null,
     line_title_id    uuid references line_title (id) not null,
-    rate             integer                         not null,
+    rate             double precision                         not null,
     created_by       uuid references app_user (id)   not null,
     updated_by       uuid references app_user (id),
     created_at       bigint default extract(epoch from now()) * 1000000,
